+++ src/client.ts
@@ -9,7 +9,7 @@ import * as logger from './logging';
 
 import { postgresqlErrorCodes } from './errors';
 import { Queue } from './queue';
-import { Query } from './query';
+import { Query, QueryParameter } from './query';
 
 import { ConnectionOptions, TLSSocket, connect as tls, createSecureContext } from 'tls';
 
@@ -588,23 +588,35 @@ export class Client {
             });
     }
 
+    /**
+     * Send a query to the database.
+     *
+     * The query string is given as the first argument, or pass a {@link QueryParameter}
+     * object which provides more control.
+     *
+     * @param text - The query string, or pass a {@link QueryParameter}
+     *     object which provides more control (including streaming values into a socket).
+     * @param values - The query parameters, corresponding to $1, $2, etc.
+     * @param types - Allows making the database native type explicit for some or all
+     *     columns.
+     * @param format - Whether column data should be transferred using text or binary mode.
+     * @param streams - A mapping from column name to a socket, e.g. an open file.
+     * @returns A promise for the query results.
+     */
     query<T = ResultRecord>(
-        text: string,
+        text: QueryParameter | string,
         values?: any[],
         types?: DataType[],
         format?: DataFormat | DataFormat[],
         streams?: Record<string, Writable>):
         ResultIterator<T> {
-        const query =
-            (typeof text === 'string') ?
-                new Query(
-                    text,
-                    values, {
-                    types: types,
-                    format: format,
-                    streams: streams,
-                }) :
-                text;
+        const query = new Query(
+            text,
+            values, {
+            types: types,
+            format: format,
+            streams: streams,
+        });
         return this.execute<T>(query);
     }